MODIFY make errors

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: jjin <jjin@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/03 18:38:38 by jjin              #+#    #+#             */
-/*   Updated: 2022/05/04 21:42:57 by jjin             ###   ########seoul.kr  */
+/*   Updated: 2022/05/20 19:03:07 by jjin             ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = -1;
 	while (++i < n)
